+++ packages/blocks/core/js/wordpress/footnotes/index.js
@@ -4,11 +4,13 @@
  * External dependencies
  */
 import type { MixedElement } from 'react';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
  */
 import { SharedBlockExtension } from '@blockera/editor';
+import { Icon } from '@blockera/icons';
 
 /**
  * Internal dependencies
@@ -20,7 +22,12 @@ export const Footnotes: BlockType = {
 	name: 'blockeraFootnotes',
 	targetBlock: 'core/footnotes',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			icon: <Icon icon="block-footnote-return" iconSize="20" />,
+			label: __('Jump to Footnote Links', 'blockera'),
+			description: __('The links to the footnote reference.', 'blockera'),
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;